feat: show multiple of as money

+++ catalyst_voices/apps/voices/lib/widgets/text_field/voices_money_field.dart
@@ -205,7 +205,7 @@ class _MoneyMath {
       case MoneyUnits.minorUnits:
         return Money(
           currency: currency,
-          minorUnits: BigInt.from(value) * currency.decimalDigitsFactor,
+          minorUnits: BigInt.from(value * currency.decimalDigitsFactor.toInt()),
         );
     }
   }
